restructuring code and fixing all tests to run with new zmp-based implementation
