+++ src/rostful_node/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .rostful_ctx import rostful_ctx
-from .rostful_node_process import RostfulNodeProcess
-from .rostful_client import RostfulClient
+from .pyros_ctx import pyros_ctx as rostful_ctx
+from .pyros_node import PyrosNode as RostfulNodeProcess
+from .pyros_client import PyrosClient as RostfulClient
 
 __all__ = [
     'rostful_ctx',
