+++ src/rostful_node/mockinterface/mockaction.py
