+++ src/rostful_node/mockinterface/mockinterface.py
