+++ src/rostful_node/mockinterface/mockmessage.py
