+++ src/rostful_node/mockinterface/mocknode.py
