+++ src/rostful_node/mockinterface/mockservice.py
