+++ src/rostful_node/mockinterface/mocktopic.py
