+++ src/rostful_node/rosinterface/__init__.py
