+++ src/rostful_node/rosinterface/action.py
