+++ src/rostful_node/rosinterface/deffile.py
