+++ src/rostful_node/rosinterface/definitions.py
