+++ src/rostful_node/rosinterface/exceptions.py
