+++ src/rostful_node/rosinterface/message_conversion.py
