+++ src/rostful_node/rosinterface/node.py
