+++ src/rostful_node/rosinterface/roconinterface/interaction.py
