+++ src/rostful_node/rosinterface/roconinterface/interaction_table.py
