+++ src/rostful_node/rosinterface/roconinterface/utils.py
