+++ src/rostful_node/rosinterface/ros_loader.py
