+++ src/rostful_node/rosinterface/service.py
