+++ src/rostful_node/rosinterface/topic.py
