+++ src/rostful_node/rosinterface/util.py
