+++ tests/rostful_node/nosetests/testPyrosClient.py
@@ -4,19 +4,21 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', '..', 'src')))
 
-from rostful_node.rostful_node_process import RostfulNodeProcess
-from rostful_node.rostful_client import RostfulClient
+from rostful_node.pyros_node import PyrosNode
+from rostful_node.pyros_client import PyrosClient
 
+import nose
 from nose.tools import assert_equal, assert_raises
+
 
 class TestRostfulClientOnMock(object):
     def setUp(self):
-        self.mockInstance = RostfulNodeProcess(mock=True)
+        self.mockInstance = PyrosNode(mock=True)
         cmd_conn = self.mockInstance.launch()
-        self.client = RostfulClient(cmd_conn)
+        self.client = PyrosClient(cmd_conn)
 
     def tearDown(self):
-        self.mockInstance.terminate()
+        self.mockInstance.shutdown()
 
     ### TOPICS ###
     def test_inject_None(self):  # injecting None is meaningless and should return false
@@ -25,6 +27,8 @@ class TestRostfulClientOnMock(object):
     def test_inject_Empty(self):
         assert self.client.topic_inject('random_topic')  # simply check if injected
 
+    #TODO : how to test strict backend with Mock ?
+    @nose.SkipTest
     def test_inject_Wrong(self):
         with assert_raises(Exception) as expt:  # TODO : be more specific
             data = 42
@@ -72,7 +76,8 @@ class TestRostfulClientOnMock(object):
         assert recv == {'first': 'first_string', 'second': 'second_string'}
 
     ### SERVICES ###
-
+    # TODO : think how to test strict backend with Mock ?
+    @nose.SkipTest
     def test_call_Wrong(self):
         with assert_raises(Exception) as expt:  # TODO : be more specific
             data = 42
